Refactor MessageInput component and enhance ChatMessage with copy functionality. Updated imports for MessageInput and improved ModelPicker's icon rendering. Removed console log from sendChatMessage function for cleaner code.
